requirements.txt
discord.py==1.7.3
GitPython==3.1.27
mysql_connector_repackaged==0.3.1
py_expression_eval==0.3.14
python-dotenv==0.20.0
